cambios

+++ src/components/Carrito/Carrito.jsx
@@ -6,7 +6,7 @@ import axios from 'axios';
 import { ENDPOINT } from '../../config/apiconfig'; 
 
 const Carro = () => {
-  const { productos } = useContext(ProductosContext);
+  const { productos } = useContext(ProductosContext); 
   const [carrito, setCarrito] = useState([]);
   const [loading, setLoading] = useState(false);
   const [metodoPago, setMetodoPago] = useState('');
@@ -43,84 +43,40 @@ const Carro = () => {
     }
   };
 
-  /*
   const eliminarProductoDelCarrito = async (productoId) => {
     const token = localStorage.getItem('token');
     const userId = localStorage.getItem('userId'); 
-  
+
     if (!token || !userId || !productoId) {
       toast.error('Token, ID de usuario o ID de producto no encontrados');
       return;
     }
-  
+
     try {
+      const productoAEliminar = productos.find((producto) => producto.id === productoId);
+
+      if (!productoAEliminar) {
+        toast.error('Producto no encontrado en el contexto');
+        return;
+      }
+
       const url = ENDPOINT.eliminarProductoCarrito(userId, productoId);
       const response = await axios.delete(url, {
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
-  
+
+      // Filtramos el carrito para eliminar el producto
       setCarrito((prevCarrito) => prevCarrito.filter((producto) => producto.producto_id !== productoId));
+
       toast.success('Producto eliminado del carrito');
     } catch (error) {
       toast.error('Error al eliminar el producto del carrito');
       console.error('Error al eliminar el producto del carrito:', error);
     }
   };
-*/
-const eliminarProductoDelCarrito = async (productoId) => {
-  const token = localStorage.getItem('token');
-  const userId = localStorage.getItem('userId'); 
 
- 
-
-  try {
-    
-    if (carrito.length === 0) {
-      const productosResponse = await axios.get(ENDPOINT.productos, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-
-      if (productosResponse.status === 200) {
-      
-        const producto = productosResponse.data.find(
-          (prod) => prod.producto_id === productoId
-        );
-
-        if (producto) {
-         
-          const url = ENDPOINT.eliminarProductoCarrito(userId, productoId);
-          const response = await axios.delete(url, {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          });
-
-          toast.success('Producto eliminado del carrito');
-        } else {
-          toast.error('Producto no encontrado');
-        }
-      }
-    } else {
-    
-      const url = ENDPOINT.eliminarProductoCarrito(userId, productoId);
-      const response = await axios.delete(url, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-
-      setCarrito((prevCarrito) => prevCarrito.filter((producto) => producto.producto_id !== productoId));
-      toast.success('Producto eliminado del carrito');
-    }
-  } catch (error) {
-    toast.error('Error al eliminar el producto del carrito');
-    console.error('Error al eliminar el producto del carrito:', error);
-  }
-};
   const calcularTotalCarrito = () => {
     return carrito.reduce((total, producto) => {
       const precio = parseFloat(producto.price) || 0;
